third_party/WebKit/LayoutTests/fast/canvas/canvas-pattern-set-transform-expected.txt

Test for supporting setTransform on canvas patterns

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pixel 20, 20 is [0, 128, 0, 255]
PASS successfullyParsed is true

TEST COMPLETE
